Separated the code into the main app and the component I2C temp. sensor API
