+++ components/temp_sensor/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "application.c"
+idf_component_register(SRCS "temp_sensor.c"
                     INCLUDE_DIRS "include")
